This is an incremental improvement of Sync BookmarkModelAssociator that contains the following changes:
1) The code that checks transaction version consistency of the
native bookmark model against Sync now remembers the result of the
check and reports it via Sync.BookmarksModelSyncStateAtNewDuplication
UMA counter when a new bookmark duplication has been detected.
This should allow us to better understand the state leading
to bookmark duplication.

2) Added a placeholder for the new optimistic bookmark association
algorithm behind a finch experiment switch.
The optimistic algorithm would used only for the case when the
transaction version stored in the native model is in sync with
the sync version. For now the optimistic algorithm is pretty
much the same as the existing conservative algorithm.
The only difference is that the sync data isn't propagated to
the native model for matching nodes. The next patch will build
on top of this and introduce the logic for chasing local
modifications leading to bookmark duplication.

BUG=456876,464907

Review URL: https://codereview.chromium.org/1133333004

Cr-Original-Commit-Position: refs/heads/master@{#331515}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 019ed5124478d523eae9fb40a567d8760f4e2ec4