Don't use field trials on the reference perf bots.
The current field trial config is used for an older build in the
case of reference. This is a problem because we are then subjecting
older builds to newer configurations that may crash.  To work
around this problem, don't add the field trials to reference builds.

Review URL: https://codereview.chromium.org/1162263005

Cr-Original-Commit-Position: refs/heads/master@{#332318}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6a540aff2dc576782e20e7c2a7a37b74edd0284