Refactor patch_orderfile.py.
patch_orderfile takes in a list of sections, and outputs an expanded list of
sections.  While this logic does require inferring symbol names, this
refactors the logic to be in terms of sections as much as possible.

BUG=460906

Review URL: https://codereview.chromium.org/1165603003

Cr-Original-Commit-Position: refs/heads/master@{#332376}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0944ad02d4faef7c51accd0d2fa2252a74f2789