Reorder windows in overview to present a consistent window order.
This change presents windows in overview in a consistent order and attempts to minimize how far they move in overview. This means that for a left/right split windows will tend to stay on the same side of the screen.

BUG=324813
TEST=WindowSelectorTest.MinimizeMovement, WindowSelectorTest.MinimizeMovementSecondDisplay, WindowSelectorTest.StableOrder

Review URL: https://codereview.chromium.org/1157423004

Cr-Original-Commit-Position: refs/heads/master@{#332459}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cff1d461591d623822f212bcb062a9252b4ce215