Fix remaining Android GN input/output errors
These are the remaining 5 mismatches between GN and Ninja's dependency tree.

The nontrivial one is that one mojo target was using copy_ex to copy a directory and specifying that directory as an input, which is not how GN or Ninja work. It ran because "html_viewer" matched a phony rule which happened to cause that directory to be built, and simultaneously made the copy_ex script recursively copy the directory.

I changed it to pass a separate parameter to the copy_ex script which it seems the Android templates do for this case. This means that it doesn't appear as a "source" to the script, but the dependency on the html_viewer binary target will ensure that the directory is built before the copy script runs, so everything will work correctly.

Review URL: https://codereview.chromium.org/1155453007

Cr-Original-Commit-Position: refs/heads/master@{#332912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 050e25f016e1dbdf8eb8fc375cc7e79748cefee9