Add usage metrics to CustomTabActivity
* Starts recording actions for custom menu items, custom action button,
find in page and open in chrome items.
* Starts recording a histogram based on the launching app given the app_ID
extra inside the intent.
* Starts recording a histogram about whether the currently visible tab is
a custom tab or not. This gets called from ProvideGeneralMetrics in the
metrics provider and can be used to slice any general metrics in being from
a custom tab or not.

Review URL: https://codereview.chromium.org/1182503003

Cr-Original-Commit-Position: refs/heads/master@{#334192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb97b9649dcbc2ad16ee1018dc44b725ea652eff