Fix `mb analyze` to work properly when called before `mb gen`.
With the recent shuffling of commands in the chromium_gn and
chromium recipes, it is now possible that `mb analyze` will
get called before `mb gen`, and that was not okay prior to
this patch. Now, we call `mb gen` from inside `mb analyze`,
in order to ensure that we have a valid output directory.

This should un-break the bots.

TBR=scottmg@chromium.org, phajdan.jr@chromium.org
BUG=481693
NOTRY=true

Review URL: https://codereview.chromium.org/1187423003.

Cr-Original-Commit-Position: refs/heads/master@{#335169}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0fd41bcddbf638bc24c1bb7ca52e9948d2ea5823