Add UMA metrics to track the source of homescreen icons on launch.
Currently, it is impossible to distinguish between launches from
homescreen icons added via app banners or the menu item. This CL adds an
additional value to the intent saved on homescreen icons. The value
corresponds to the source of the icon: whether it was added via the add
to homescreen menu item, via an app install, banner, or from an unknown
source (all intents added prior to this CL). The additional metric
allows for the effectiveness of web app install banners on Android to be
measured.

BUG=486177

Review URL: https://codereview.chromium.org/1220813010

Cr-Original-Commit-Position: refs/heads/master@{#337701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32fe3048050916028ac902abd0d7db9de4263e0b