Attempt to fix a crash in NerErrorHelperCore.
It looks like we're getting button press events when
committed_error_page_info_ is NULL.  This CL checks for that case,
to prevent the crash.  As it's unclear what's causing this, it also
adds a pair of histograms to help investigate the cause of the
underlying issue.

BUG=500556

Review URL: https://codereview.chromium.org/1180693006

Cr-Original-Commit-Position: refs/heads/master@{#337852}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 077f450b90cca12652637b0fdb55f3745fc74d0e