metrics: Add dbus interface for GetRandomPerfOutput
GetRandomPerfOutput is a new dbus method call that replaces the older
GetRichPerfData method call to debugd. The difference is that the new
method call can return either a PerfDataProto or a PerfStatProto.

Also updating PerfProvider to use the new dbus interface and parse and
store the result as either protobuf format.

BUG=chromium:478289
TEST=Set |kPerfProfilingIntervalMs| to something short (e.g. 15 sec)
and add logging to ParseOutputProtoIfValid() to verify that valid
PerfDataProto or PerfStatProto data was received.

Signed-off-by: Simon Que <sque@chromium.org>

Review URL: https://codereview.chromium.org/1218583002

Cr-Original-Commit-Position: refs/heads/master@{#337910}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0260fd891172eb94a614751b1c77596e476bab2