Generate a static struct from fieldtrial_testing_config
This is gives Chromium developers visibility into what features are active, but hidden behind field trial testing while developing Chromium.

There is already a JSON file in Chromium with which field trials should be active. The goal is make a code generator to output a struct based off the contents of this JSON file. The next step after this is to create a mechanism to activate the trials defined by this generated struct.

BUG=504075

Review URL: https://codereview.chromium.org/1209743002

Cr-Original-Commit-Position: refs/heads/master@{#338416}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 970b79a8adc8c3fa3593054681b6d4bc5ae3ec94