[telemetry] Service noise benchmark.
A benchmark intended to run on a Windows bot. It detects the amount of noise that various Windows services contribute to our regular perf benchmarks. As a proxy for noise, we use power consumption, because that metric is most sensitive to activity.

It starts by measures the power usage of the system (called default).
Then it disables all the suspicious services and measures it again (called control).
Then it enables each service in turn and measures it again.

BUG=416072

Review URL: https://codereview.chromium.org/1223833007

Cr-Original-Commit-Position: refs/heads/master@{#338605}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91f2d7c36f71b7a115a0ce23977f73618ab83ef0