Introduce the PushSubscription.curve25519dh attribute (2/3)
This CL provides plumbing for sending the curve25519dh attribute to Blink
in the appropriate places. Hooking it up with the GCM Driver will be
done in a later patch.

This CL is part of a three-sided patch:
  [1] https://codereview.chromium.org/1222903007
  [2] This CL.
  [3] https://codereview.chromium.org/1237673002

BUG=486040

Review URL: https://codereview.chromium.org/1210283005

Cr-Original-Commit-Position: refs/heads/master@{#338702}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb008005a0fe6ec2fd0865732f5e4a7b89f08c8f