Make android sign in promo UMA metric visible on dashboard.
The code that adds the UMA logging:
https://chrome-internal-review.googlesource.com/#/c/175857/22

is not consistent with the change in the histograms.xml file:
https://codereview.chromium.org/568603002/

which means the UMA dashboard does not show the metric.  This CL changes the
histograms.xml to match the name of the metric in the code.

BUG=None

Review URL: https://codereview.chromium.org/1257383002

Cr-Original-Commit-Position: refs/heads/master@{#340738}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 976c88b9e003cad1d481660286e5e825b7e568f0