Add UMA to track app install banner outcomes.
It is difficult to get a picture for what is happening when banners are
shown, as three separate metrics track a number of fine-grained
outcomes.

This CL adds a new UMA solely tracking the user response to a banner,
that is, once a banner is shown, whether users accept, dismiss, or
ignore it. Native and web apps are tracked by separate metrics.

BUG=516956

Review URL: https://codereview.chromium.org/1269303002

Cr-Original-Commit-Position: refs/heads/master@{#342270}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1bd97d30773b89a6be3c07eb7f85ae284e82462d