Add Precache.Latency.Overall metric.
This metric records the response latency for requests that don't occur
as part of precaching. The metric is computed in
ChromeNetworkDelegate::OnCompleted and logged in
PrecacheDatabase::RecordURLFetched.

Also, move HistoryService initialization to PrecacheManagerFactory. This is just cleanup now, but will be used in a future commit.

BUG=499532

Review URL: https://codereview.chromium.org/1278693002

Cr-Original-Commit-Position: refs/heads/master@{#342856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 173e75198927528a8f7fb786ad8c93855f658ab0