Add {,Non}TopHosts variants of Precache.Latency.
Add Precache.Latency.NonPrefetch.TopHosts, which records the latency
only for non-prefetch requests whose referer is in the top hosts, and
Precache.Latency.NonPrefetch.NonTopHosts, which records the latency for
the rest of the non-prefetch requests. This allows us to evaluate the
effect of precaching on latency only on those requests it has a chance
of improving. Thus, we can better distinguish the effectiveness of the
host selection algorithm from the effectiveness of the manifests.

This adds a HostRankIfAvailable method to HistoryService, in order to expose
the pre-computed TopHosts, as requiring that computation for every request
would be untenable.

This also adds a Control group, where TopHosts is computed for users, but no
resources are prefetched.

BUG=499532

Review URL: https://codereview.chromium.org/1280673004

Cr-Original-Commit-Position: refs/heads/master@{#343230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 981c7eacd300a5bfb885a120e642e329a4eb68ce