[Android] Add UMA histograms to measure context menu actions.
This adds four related histograms to measure how users interact with
context menus. The histograms are ContextMenu.Link, ContextMenu.Image,
ContextMenu.ImageLink, and ContextMenu.Video, and the values are the
various actions that can be taken on each context menu, e.g. "Open in
new tab" or "Save image".

For example, if a user long presses an image and selects "Copy image",
we'll record the "Copy image" event in the ContextMenu.Image histogram.

ContextMenu.ImageLink is used when the user long presses an image which
is also a link. This case is treated specially because its context menu
is currently so long that it needs extra love and investigation.

BUG=483685

Review URL: https://codereview.chromium.org/1267213002

Cr-Original-Commit-Position: refs/heads/master@{#343283}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a60a2e9061334a5726ae30166ef004209719f4f9