Don't reload twice in key_mobile_sites
The Linkedin and WoW wiki pages in key_mobile_sites reload themselves
before scrolling to take advantage of the browser's shader cache.
However this reload is currently implemented as a reload plus a
navigation. Since the navigation has more or less the same effect as a
reload, the reload step is redundant.

This patch removes the explicit reload step. This will make the
benchmark slightly faster, and with any luck, make it less flaky[1] on
some downstream bots.

Note that replacing the second navigation with just the reload doesn't
work because the reload does not reset the scroll offset, leaving us at
the bottom of the page.

[1] https://go/lhivi

Review URL: https://codereview.chromium.org/1296223002

Cr-Original-Commit-Position: refs/heads/master@{#343690}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6416d003916e852c9adf2f69d7cbe6582763c7fa