[Telemetry] Remove --chartjson flag
The --chartjson flag is useless, unused, and misleading. There are a bunch of
apocryphal reasons why the flag was created, but in the churn of switching to
the output formatter abstraction last year this flag unfortunately didn't get
removed when it should have.

Review URL: https://codereview.chromium.org/1305163003

Cr-Original-Commit-Position: refs/heads/master@{#345153}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9a1e140a517d24a1e081545ffa0570a77ab6ef9