Count the number of clicks users make on Site Settings buttons.
This adds an action to be counted when a user clicks on a [Site Settings]
button included with a Web Notifications. Right now we only track this
on Android, but this will be counted on desktop as well when the button
gets added (crbug.com/524474).

R=miguelg
BUG=

Review URL: https://codereview.chromium.org/1315873002

Cr-Original-Commit-Position: refs/heads/master@{#346137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e49580721010d46b84483898a1d721482a4bcd8