[Telemetry] Introduce SummarizableValue.
This is a refresh of crrev.com/809393002, wherein
improvement_direction was added to Value.

This CL pulls out only the Telemetry-side changes
in an attempt to make this CL a bit smaller; the
next change will be to update Telemetry clients to
use this new field.

BUG=459450

Review URL: https://codereview.chromium.org/1313243003

Cr-Original-Commit-Position: refs/heads/master@{#346410}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ad6f2b91267a69d173713e1607774cac8bbf58a