Enable background mode if there is at least one Push API subscription.
This introduces the concept of the BackgroundTrigger. If the BackgroundModeManager has at least one trigger, or a background app (extension), it needs to run.

This is disabled by default, and can be enabled with the enable-push-api-background-mode flag, or in chrome:flags.

This is only enabled on Windows and Linux, as discussed in the bug.

BUG=402456

Review URL: https://codereview.chromium.org/1124263002

Cr-Original-Commit-Position: refs/heads/master@{#346660}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5abba42cdc79bcc9038338678ec49236de229608