Add a flag to enable testing of WebUSB by bypassing the permssions UI.
The permissions UI for WebUSB is not yet implemented. This flag bypasses
the need for a site to call navigator.usb.requestDevice to prompt the
user for access to USB device before calling navigator.usb.getDevices.
As such it is a Bad Flag(TM) but the risk is still rather low because
the device is still required to have a set of WebUSB descriptors
granting the site's origin access.

BUG=529950

Review URL: https://codereview.chromium.org/1333703002

Cr-Original-Commit-Position: refs/heads/master@{#348091}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9af433a36624bbb53800c933c8027a584ca5ad0