Make MB clobber build directories when switching between GYP and GN.
GYP and GN put generated files in different places inside the build
directory. This means that if you do a GYP build and then a GN build
into the same directory, you may get unpredictable results depending
on which generated files are picked up at what times.

In order to avoid this, we modify MB to keep track of whether a
given build directory has build files generated by GYP or GN; if there
is a mismatch (like when we flip a bot from GYP to GN), we clobber the
build directory and start from scratch to be safe.

In addition, the first time we enable MB on a bot, we will not know
what the existing build directory contains; to be safe, we do a clobber
in this situation as well.

R=brettw@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1338123002

Cr-Original-Commit-Position: refs/heads/master@{#348705}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c161aa945c811b260f7f2981e592f63d16c3a106