Clean up logging in MB.
By default, the bots run `mb -v`, which dumps a lot of output,
especially when generating isolates for GN. This patch cleans
up a lot of the output so that we log a minimal subset most of the
time and more on errors (but -v will still dump everything).

In the common (non-error) case, this means that we'll get only 1-2
lines of output per MB invocation.

In addition, this patch removes the --quiet flag (which was never used)
and the RunGNIsolate() function, which was never called, and fixes
a minor but annoying issue where we would list the same dependency
multiple times in the analyze results.

R=thakis@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1342563002

Cr-Original-Commit-Position: refs/heads/master@{#348806}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0547cd084409dd5f6c336a2babb47ca28a66d1f