Remove the 'gyp_config' concept from MB.
Previously MB would try to ensure that the build directory
passed to it matched the expected gyp configuration for the bot
(i.e. that we would not pass "//out/Debug" as a path for a
release bot).

Making this work correctly was awkward for the Win x64 bots
(which use Release_x64 instead of Release), and it's not clear
that this check would ever actually catch a real problem.

So, this patch deletes all of the 'gyp_config' logic instead
and simplifies things.

TBR=scottmg@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1348463002

Cr-Original-Commit-Position: refs/heads/master@{#349043}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fae80e312dba6818315f4f8f0c3a1d85f6f4f9fb