Switch from using codec ids to codec name hashes.
Uses 32-bits of the SHA-1 hash of the codec name as the sparse
histogram identifier.  Adds a unittest for producing this enum
mapping so it can be updated easily.

FFmpeg is always changing codec ids, so using the raw AVCodecID
is proving to be untenable.

BUG=none
TEST=playback shows the right hashes.

Review URL: https://codereview.chromium.org/1348563002

Cr-Original-Commit-Position: refs/heads/master@{#349472}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d1db5e968f76dc92145cbcf49abda7cbf1bbc0f