Exposing a new BLE connection API in the EasyUnlock private JS API.
The connection manager observe the connection events and dispatches them
to the JS API. The BLE connections are managed using the ApiResourceManager.

The EasyUnlockPrivateConnection is a wrapper for a
proximity_auth::Connection, and its lifetime is managed by the
ApiResourceManager.

BUG=528274

Review URL: https://codereview.chromium.org/1333343002

Cr-Original-Commit-Position: refs/heads/master@{#349650}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d303ae9785a772691c750aa254dd5db7fec156c