PlzNavigate: do not add ResourceThrottles for blob urls
With PlzNavigate enabled, the renderer will request a blob url to access the
body of a request already fetched by the browser. This CL makes it so that the
ResourceThrottles will not be added for those requests, since policy decisions
should have been made on the initial request. In particular, this prevents the
PrerenderContents from being canceled by the PrerenderResourceThrottle.

BUG=504347

Review URL: https://codereview.chromium.org/1348863002

Cr-Original-Commit-Position: refs/heads/master@{#349972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f468a1cd74613d40da3ee113750dd60e2fb85c9