[Telemetry] Parameterize merging and summary by a key function.
This change adds the ability to pass a nonstandard key_func to Summary and merge_values's public interfaces (MergeLikeValuesFromSamePage and MergeLikeValuesFromDifferentPages).

Crucially, it also implements merge_values.DefaultKeyFunc and specifies it by default in merge_values and summary. This means that when the default grouping for values changes as it will for TBM, all we will have to do is modify the implementation of merge_values.DefaultKeyFunc.

BUG=534909

Review URL: https://codereview.chromium.org/1363653002

Cr-Original-Commit-Position: refs/heads/master@{#350640}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62ab9d333f5eb071bbd314491cfe7071dd332921