Exposing a new easyUnlock.setupConnectionGetDeviceAddress JS API.
This CL exposes a new JS API to find the bluetooth address for the
remote device corresponding to a given connection id.

This is necesssary as the bluetooth address of a device changes after
it's paired (but it's updated in the underlying
BluetoothLowEnergyConnection object).

BUG=528277

Review URL: https://codereview.chromium.org/1370133005

Cr-Original-Commit-Position: refs/heads/master@{#351549}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4ca7e269f215e183571f3025de7570aac92e379